requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
fastapi==0.115.0
fastavro==1.9.7
huggingface-hub==0.25.1
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.50.2
pandas==2.2.3
pdfminer.six==20240706
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
requests-mock==1.12.1
scikit-learn==1.5.2
scipy==1.14.1
semantic-chunkers==0.0.10
semantic-router==0.0.68
uvicorn==0.31.0
python-telegram-bot==21.6
python-dotenv==1.0.1